remove the feature flag and reorganize the module slightly
