+++ meilisearch/src/analytics/segment_analytics.rs
@@ -255,8 +255,7 @@ impl From<Opt> for Infos {
         // to add analytics when we add a field in the Opt.
         // Thus we must not insert `..` at the end.
         let Opt {
-            #[cfg(features = "metrics")]
-                enable_metrics_route: _,
+            enable_metrics_route: _,
             db_path,
             http_addr,
             master_key: _,
